+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added support for anonymous connections to MQTT brokers
+
 ## [1.0.1] - 2023-12-28
 
 ### Fixed