+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2025-06-03
+
 ### Added
 
 - Added parameter for the retain property for Send-MQTTMessage (-Retain) (defaults to false)